Add tc_new.h, tc_initializer_list.h.
tc_new.h avoids our needing to #include<new> on platforms that lack the
standard C++ library, but also does not conflict with it if that is
included.

tc_initializer_list.h is a replacement for <initializer_list> to be used
only on platforms that lack this C++ library header (AVR).